pyproject.toml
[project]
name = "verity"
version = "0.1.0"
description = "AI Compliance Assistant"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "langchain-aws>=1.1.1",
    "langchain-google-genai>=4.1.2",

    "google-genai>=1.56.0",

    "faiss-cpu>=1.8",
    "opensearch-py>=2.5",
    "requests-aws4auth>=1.3",
    "boto3>=1.34",

    "pandas>=2.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "pydeck>=0.9",

    "streamlit>=1.30",
    "streamlit-elements>=0.1",
    "streamlit-shadcn-ui>=0.1",

    "pypdf>=3.17",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "python-dotenv>=1.0",
    "chardet>=5.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "pandas-stubs",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PL"]
ignore = ["PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
